Remember the user's date range (OCEEBMS-651)
The users had earlier asked that we set the default for the date
range on the Import Report request form to empty. At first we took
that request at face value, setting the Drupal FAPI #default_value
property to an empty string for each of the two date fields. But
what the users really wanted was for us to do that only when the
user first brings up the request form. When the user submits the
request and we display the report, the form should remember the
date range the user just specified. So now we go back and get
those values and use them for the #default_value properties when
the form is redrawn.