Fix search title generation (OCEEBMS-684)
We need to use dom_import_simple_xml() to get a DOM node for the
ArticleTitle element, so that we can use its textContent property.
That's the only way we can reliably convert the character entities
back to the characters they represent, in order for sorting and
searching to work correctly.