Change order of decision buttons (OCEEBMS-698)
This makes the decision-handling code tricker, because the queues
no longer share all the decisions at the front of the a common
decisions array (this ticket jumbled the order of the decision
radio buttons). The keys for the DECISIONS array represent the
position in the array of decision buttons shown for each
article/topic combination for the selected queue (with zero
being the position of the "None" button, which never triggers
storage of a new state).