Fix full article history crash (OCEEBMS-718)
Article 730646 is one of the six duplicate pairs, sharing PMID
31749144 with article 730647 (see OCEEBMS-717). The original
EBMS didn't anticipate such duplicates, so the refresh code only
found one of the articles in the pair when it retrieved updated
XML from PubMed. But each time the refresh job ran, it noticed
that the other article of the pair needed to be refreshed, so it
pulled down the XML again, but only refreshed the first article.
So at the time the problem was noticed, article 730646 had 1,031
import refresh jobs associated with it, and the attempt to load
the Batch entities for all of those jobs caused the PHP interpreter
to run out of memory. So we have switched this query to use the
Drupal Database API.