Work around bugs in NLM APIs (OCEEBMS-687, OCEEBMS-719)
We discovered that NLM's APIs do not reveal all of the PubMed
articles which have changed. As a result the production EBMS
has thousands of articles whose XML is out of date. We filed
a ticket reporting the problem, but they do not appear willing
to correct it. So we have rewritten the nightly job which
refreshes the article XML, as well as the software preparing
for the migration to the new Drupal 9 EBMS.