Use data_checked for article XML refresh logic
Set this field every time the article's XML is fetched and saved
in the EBMS, even if the data we store hasn't changed. This is in
contrast with the update_date field, which is only set if the data
has changed. The scheduled refresh job only queues up an article
for refresh is its DateRevised is on the same day or later than
the latest of the article's EBMS import_date, update_date, and
data_checked dates.