Clear the cache before loading articles
The migration script is still crashing on the CBIIT servers.
The latest database errors indicate that the crash happend when
Drupal is trying to update the cache tables. Also interesting,
the migration crashed on both the DEV and QA servers after
loading same article. I was able to remove that article and
run a hacked version of the article loading script (modified
to skip over the articles which had already been loaded) and
that job was able to finish loading all of the articles, which
would seem to rule out the possibility that something about
one of the articles triggers intrusion detection software on
the servers or network. So my best theory is that the cache
reaches a certain size which exceeds the capacity of the
systems to handle whatever Drupal needs to do. So now, right
before invoking the script to load the articles the master
script sleeps for a minute (giving the systems a chance to
catch their breath and do whatever housekeeping the need to
do), clears the Drupal caches, and sleeps for another minute.