Split the migration build script into two parts
On some of the CBIIT servers, the full migration build script
is never able to run to completion successfully, at some point
failing with the error message saying that "MySQL server has
gone away" (this has never happened on my own server, leading
to the conclusion that this is more likely an environment problem
rather than a problem with the software). So I have split the
full script into two parts, to be run on separate days (to avoid
the scheduled maintenance which runs at midnight and kills the
job). The most failures have been encountered on the QA server.