Fix event navigation edge case (OCEEBMS-733)
While working on OCEEBMS-732 I stumbled on a rare condition in
which the event navigation buttons ("Prev" and "Next") would
cycle back and forth between the same two events. This only
happens when multiple events which are all visible for the
current user are scheduled to start at exactly the same time on
the same day. I had ended up with this condition as a result of
creating multiple meetings in rapid succession for testing, using
the default settings. The condition is very much an edge case (as
a general rule, I imagine booking the same board member for
multiple meetings at the same time would be discouraged), and I
was strongly tempted to ignore the problem, as handling this edge
case correctly would double the number of database queries used
to create these buttons. But in the end I couldn't bring myself
to leave this behavior in place, so I'm fixing it.