Fix the #upload_location path (OCEEBMS-734)
Although (as far as I can tell) the Drupal documentation doesn't
say so, while the #upload_location property takes a trailing slash
for a path with a subdirectory, that extra slash causes random
failures if there is no subdirectory in the path. So 'public:///'
fails (sometimes, but mysteriously not all the time) whereas
'public://' works (all the time). This is what was causing posting
of files to the Meeting form to blow up.