Split migration build into three parts
It was not sufficient to break the migration build into two
separate parts for the CBIIT servers, as the second part was
consistently failing in the middle of loading the article
entities. Therefore, the articles.php script has itself been
split into two parts, and now migrate-part2.sh runs only the
first of those, with migrate-part3.sh finishing the migration
steps.