Remove core journal sort (OCEEBMS-743)
Sorting by core journal is so slow that with a large queue the
callback to record decisions can still be pending when the user
clicks the button to apply the decisions she has made. Now that
we have removed the display of pending decisions, the user has
no way to detect this condition when clicking that button.
A checkbox has been added to filter to only show articles which
have been published in core journals, to replace the functionality
provided by core journal sorting.