Fix publishing queue checkboxes (OCEEBMS-759)
We're reasonably confident that the checkboxes on the Publish
Articles page were working when they were first implemented, and
when they were tested during user acceptance testing prior to
launch. If that's true, then a last-minute Drupal upgrade must
have modified the ajax behavior (perhaps fixing a bug we were
working around). Two changes were required. The first was to
replace the direct binding with .click() in the Javascript to
event delegation using .on().[1,2] The second was to make sure
we get the value of the hidden queued field using getUserInput()
instead of from the loaded parameters.

[1] https://stackoverflow.com/questions/21239210
[2] https://stackoverflow.com/questions/16062899