Lower default batch size (OCEEBMS-782)
Although PubMed's documentation claims that it returns the first
10,000 PMIDs in response to search queries, if you try to use that
number of PMIDs the results are truncated. This causes the report to
find articles which NLM has lost to show false positives (I suppose
you could call them "false negatives" in this context, since they
appear to report articles as having disappeared, when in fact NLM
still has them in the database). Lowering the default to 9,000 avoids
this bug.