Eliminate annoying warnings on the status page
Normally we wouldn't want to use a version of Ubuntu which is not
an LTS release, but that was the only way we could eliminate the
bogus warning about an OPcache bug in PHP 8.1 which Ubuntu 22.04
fixed by patching the bug instead of upgrading PHP to a version
which does is known to have fixed the bug. See Drupal core issue
https://www.drupal.org/project/drupal/issues/3307248.

Also elimated the warning about using a non-preferred database
transaction isolation level setting.