Avoid ugly deprecation message
All of a sudden, opening a zip archive with the CREATE option
using a path generated using tempnam() is no longer acceptable.
This is because, as it turns out, tempnam() doesn't just create
a unique name string, it also creates a file at that location.
Isn't PHP just full of wonderful surprises?