Fix Drupal silliness
With the help of an AI bot, I was able to figure out the very
obsccure PHP code which was needed to eliminate the unhelpful
error message appearing on the status page complaining that there
were "mismatched entity and/or field definitions" which updatedb
wasn't able to fix. Didn't get a response when I reached out on
Drupal's Slack support channel, but doing some digging, I found
that this is a long-standing problem, and to Drupal's shame, the
Drupal maintainers see no problem with users having to deal with
these errors.